requirements.txt
gTTS
openai
playsound3
pydub
PyAudio
python-dotenv
SpeechRecognition
ttkbootstrap
